requirements.txt
coloredlogs
flask
grpcio
grpcio-tools
pyaes
pendulum